restructure pipelines
