+++ smse/pipelines/__init__.py
@@ -1,0 +1,5 @@
+from smse.pipelines.base import BasePipeline
+from smse.pipelines.audio import AudioPipeline
+from smse.pipelines.image import ImagePipeline
+from smse.pipelines.text import TextPipeline
+from smse.pipelines.video import VideoPipeline
