+++ smse/pipelines/base.py
